Expose default device ID via DeviceCache and Authsignal.getDeviceID()
